yolo_exp/requirements.txt
opencv-python
numpy
requests
pyserial
python-dotenv
pyttsx3
SpeechRecognition
